python3/requirements.txt
pathlib==1.0.1
PyYAML==5.3.1